requirements.txt
aiohttp==3.8.1
aioitertools==0.9.0
aiosignal==1.2.0
annotated-types==0.7.0
ansible-base==2.10.8
anyio==4.6.0
apache-libcloud==3.2.0
argcomplete==1.8.1
asgiref==3.8.1
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
boto3==1.35.65
botocore==1.35.65
certifi==2021.10.8
cffi==1.17.1
chardet==3.0.4
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
Deprecated==1.2.15
distro==1.7.0
dnspython==2.1.0
docutils==0.18.1
exceptiongroup==1.2.2
fastapi==0.115.5
filelock==3.0.12
frozenlist==1.3.0
googleapis-common-protos==1.56.1
greenlet==1.1.2
grpcio==1.68.0
h11==0.14.0
httplib2==0.20.2
hyperlink==21.0.0
idna==2.8
importlib_metadata==8.4.0
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.0.3
jmespath==0.10.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
lark==1.1.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
localstack-client==2.2
lockfile==0.12.2
MarkupSafe==2.0.1
more-itertools==8.10.0
multidict==6.0.2
netaddr==0.8.0
netifaces==0.11.0
ntlm-auth==1.4.0
oauthlib==3.2.0
opentelemetry-api==1.28.2
opentelemetry-exporter-jaeger==1.17.0
opentelemetry-exporter-jaeger-proto-grpc==1.17.0
opentelemetry-exporter-jaeger-thrift==1.17.0
opentelemetry-exporter-otlp==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-instrumentation-requests==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
packaging==21.3
protobuf==5.28.3
psycopg2==2.9.10
pyaml==21.10.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.27.0
pycparser==2.22
pycryptodomex==3.11.0
pycurl==7.44.1
pydantic==2.9.2
pydantic_core==2.23.4
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pykerberos==1.1.14
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-daemon==2.2.4
python-dateutil==2.8.2
python-hcl2==4.3.2
pytz==2021.3
pywinrm==0.3.0
PyYAML==6.0
requests==2.22.0
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-toolbelt==0.9.1
rethinkdb==2.4.8
rsa==4.7.2
s3transfer==0.10.0
SecretStorage==3.3.1
service-identity==18.1.0
simplejson==3.17.6
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
systemd-python==234
terraform-local==0.11
thrift==0.21.0
Twisted==22.1.0
typing_extensions==4.12.2
urllib3==1.25.11
uvicorn==0.32.1
wadllib==1.3.6
wrapt==1.13.3
xmltodict==0.12.0
yarl==1.7.2
zipp==1.0.0
zope.interface==5.4.0